First major commit for FED signup form design.
This is a first draft, enough to move on with development. Need to
investigate field masking, client and server validation and whether the
material design buttons are the best choice. The CSS has all been
mashed together, lots of overriding styles, but I will sort it properly
when I move to bootstrap for responsive framework.